Big refactor to allow to load adapter-states
