+++ nemo/collections/llm/peft/api.py
@@ -1,4 +1,4 @@
-from nemo.lightning.peft import PEFT
+from nemo.lightning.pytorch.callbacks.peft import PEFT
 from nemo.collections.llm.peft.lora import LoRA
 from nemo.collections.llm.utils import factory
