--- nemo/lightning/peft.py
@@ -1,184 +1,0 @@
-from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Optional, Tuple
-
-import torch.nn as nn
-from nemo.lightning.io.mixin import IOMixin
-
-
-if TYPE_CHECKING:
-    from megatron.core.dist_checkpointing.mapping import ShardedStateDict
-
-
-class AdapterWrapper(nn.Module):
-    """Abstract base class for wrapping modules with adapters in Parameter-Efficient Fine-Tuning (PEFT).
-
-    This class wraps a module and its associated adapter, providing methods for
-    managing the state dictionaries of both the main module and the adapter. It does not
-    implement the forward method, which must be implemented by concrete subclasses.
-
-    Attributes:
-        to_wrap (nn.Module): The main module to be wrapped.
-        adapter (nn.Module): The adapter module to be applied.
-
-    Note:
-        This class is abstract and cannot be instantiated directly. Subclasses must
-        implement the forward method.
-
-    Example:
-        class AdapterParallelAdd(AdapterWrapper):
-            def __init__(self, to_wrap, adapter):
-                super().__init__(to_wrap, adapter)
-
-            def forward(self, x):
-                return self.to_wrap(x) + self.adapter(x)
-
-        main_module = nn.Linear(100, 100)
-        adapter = nn.Linear(100, 100)
-        parallel_adapter = AdapterParallelAdd(main_module, adapter)
-    """
-
-    def __init__(self, to_wrap: nn.Module, adapter: nn.Module):
-        super(AdapterWrapper, self).__init__()
-        self.to_wrap = to_wrap
-        self.adapter = adapter
-
-    def state_dict(self, destination=None, prefix='', keep_vars=False):
-        """Retrieve the state dictionary of the wrapped module and adapter.
-
-        This method overrides the default state_dict behavior to include both
-        the main module's state and the adapter's state under a special 'adapters' key.
-
-        Args:
-            destination (Optional[dict]): A dictionary to store the state. If None, a new
-                                          dictionary is created. Defaults to None.
-            prefix (str): A prefix added to parameter and buffer names. Defaults to ''.
-            keep_vars (bool): If True, returns variables instead of tensor values.
-                              Defaults to False.
-
-        Returns:
-            dict: The state dictionary containing both the main module and adapter states.
-        """
-
-        if destination is None:
-            destination = {}
-
-        # Get state dict of the main module
-        main_state_dict = self.to_wrap.state_dict(destination, prefix, keep_vars)
-
-        # Store adapter state dict under the special "adapters" key in the destination dict
-        adapter_state_dict = self.adapter.state_dict(None, prefix, keep_vars)
-        destination[f'{prefix}adapters'] = adapter_state_dict
-        return main_state_dict
-
-    def sharded_state_dict(
-        self,
-        prefix: str = '',
-        sharded_offsets: Tuple[Tuple[int, int, int]] = (),
-        metadata: Optional[dict] = None,
-    ) -> "ShardedStateDict":
-        """Retrieve the sharded state dictionary of the wrapped module and adapter.
-
-        This method is used for distributed checkpointing, combining the sharded states
-        of both the main module and the adapter.
-
-        Args:
-            prefix (str): A prefix added to parameter and buffer names. Defaults to ''.
-            sharded_offsets (Tuple[Tuple[int, int, int]]): Offsets for sharded parameters.
-                                                           Defaults to an empty tuple.
-            metadata (Optional[dict]): Additional metadata for the sharded state.
-                                       Defaults to None.
-
-        Returns:
-            ShardedStateDict: The combined sharded state dictionary.
-        """
-        sharded_state_dict = {}
-        sharded_state_dict.update(self.to_wrap.sharded_state_dict(prefix, sharded_offsets, metadata))
-        sharded_state_dict.update(self.adapter.sharded_state_dict(f"{prefix}adapter.", sharded_offsets, metadata))
-        return sharded_state_dict
-
-    def load_state_dict(self, state_dict, strict=True):
-        """Load a state dictionary into the wrapped module and adapter.
-
-        This method overrides the default load_state_dict behavior to handle
-        loading states for both the main module and the adapter.
-
-        Args:
-            state_dict (dict): The state dictionary to load.
-            strict (bool): Whether to strictly enforce that the keys in state_dict
-                           match the keys returned by this module's state_dict()
-                           function. Defaults to True.
-        """
-        # Check if the 'adapters' key is present in the state_dict
-        if 'adapters' in state_dict:
-            adapter_state_dict = state_dict.pop('adapters')
-        else:
-            adapter_state_dict = {}
-
-        # Load the main module state dict
-        self.to_wrap.load_state_dict(state_dict, strict)
-
-        # Load the adapter module state dict if present
-        if adapter_state_dict:
-            self.adapter.load_state_dict(adapter_state_dict, strict)
-
-
-class PEFT(ABC, IOMixin):
-    """Abstract base class for Parameter-Efficient Fine-Tuning (PEFT) methods.
-
-    This class defines the interface for PEFT methods, which are used to fine-tune
-    large language models efficiently by modifying only a small subset of the model's
-    parameters.
-
-    Example:
-        class MyPEFT(PEFT):
-            def transform(self, module, name=None, prefix=None):
-                # Implement the transform logic
-                pass
-
-
-        peft = MyPEFT()
-        peft_model = LargeLanguageModel(model_transform=peft)
-    """
-
-    @abstractmethod
-    def transform(self, module, name=None, prefix=None):
-        """Transform a single module according to the PEFT method.
-
-        This method is called for each module in the model during the PEFT application process.
-        It should be implemented by subclasses to define how individual modules are transformed
-        for the specific PEFT technique.
-
-        Args:
-            module (nn.Module): The individual module to be transformed.
-            name (Optional[str]): The name of the module within the model structure. Defaults to None.
-            prefix (Optional[str]): A prefix to be added to the module name, typically used for
-                                    nested modules. Defaults to None.
-
-        Returns:
-            nn.Module: The transformed module. This can be the original module with modifications,
-                       a new module replacing the original, or the original module if no
-                       transformation is needed for this specific module.
-
-        Note:
-            This method is automatically called for each module in the model when the PEFT
-            instance is applied to the model using the __call__ method.
-        """
-        raise NotImplementedError("The transform method should be implemented by subclasses.")
-
-    def __call__(self, model: nn.Module) -> nn.Module:
-        """Apply the PEFT method to the entire model.
-
-        This method freezes the model parameters and walks through the model
-        structure, applying the transform method to each module.
-
-        Args:
-            model (nn.Module): The model to be fine-tuned.
-
-        Returns:
-            nn.Module: The transformed model with PEFT applied.
-        """
-
-        model.freeze()
-        model.walk(self.transform)
-
-        return model
