+++ nemo/lightning/pytorch/callbacks/__init__.py
@@ -1,5 +1,6 @@
 from lightning.pytorch.callbacks.model_checkpoint import ModelCheckpoint
 from nemo.lightning.pytorch.callbacks.model_transform import ModelTransform
+from nemo.lightning.pytorch.callbacks.peft import PEFT
 from nemo.lightning.pytorch.callbacks.progress import MegatronProgressBar
 
-__all__ = ["MegatronProgressBar", "ModelCheckpoint", "ModelTransform"]
+__all__ = ["MegatronProgressBar", "ModelCheckpoint", "ModelTransform", "PEFT"]
