+++ nemo/lightning/pytorch/strategies.py
@@ -542,8 +542,6 @@ class MegatronStrategy(DDPStrategy, io.IOMixin):
     def checkpoint_io(self) -> CheckpointIO:
         if self._checkpoint_io is None:
             self._checkpoint_io = MegatronCheckpointIO()
-        elif isinstance(self._checkpoint_io, _WrappingCheckpointIO):
-            self._checkpoint_io.checkpoint_io = MegatronCheckpointIO()
 
         return self._checkpoint_io
